V2.0.0

+++ flowjob-broker-core/src/main/java/org/limbo/flowjob/broker/core/worker/metric/WorkerAvailableResource.java
@@ -1,24 +1,26 @@
 /*
- * Copyright 2020-2024 Limbo Team (https://github.com/limbo-world).
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ *  * Copyright 2020-2024 Limbo Team (https://github.com/limbo-world).
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License");
+ *  * you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  * 	http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the License for the specific language governing permissions and
+ *  * limitations under the License.
  *
- * 	http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
  */
 
 package org.limbo.flowjob.broker.core.worker.metric;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
-import org.limbo.flowjob.api.remote.param.WorkerResourceParam;
+import org.limbo.flowjob.api.param.broker.WorkerResourceParam;
 
 /**
  * worker节点上可用的资源，资源有以下三种定义：内存、CPU。值对象。
@@ -35,9 +37,9 @@ public class WorkerAvailableResource {
     private float availableCpu;
 
     /**
-     * 可用的内存空间，单位GB。
+     * 可用的内存空间，单位MB。
      */
-    private float availableRam;
+    private long availableRam;
 
     /**
      * 任务队列剩余可排队数
@@ -46,7 +48,7 @@ public class WorkerAvailableResource {
 
     public WorkerAvailableResource(
             @JsonProperty("availableCpu") float availableCpu,
-            @JsonProperty("availableRam") float availableRam,
+            @JsonProperty("availableRam") long availableRam,
             @JsonProperty("availableQueueLimit") int availableQueueLimit) {
         this.availableCpu = availableCpu;
         this.availableRam = availableRam;
